Modified Frame.rb
The changes were base on rubocop warnings.
next_frame method:
simplified and add a statent were stop add if the frame reaches the 10
next_turn:
instead of add a element in the rolls instance variable, the element is
already initialized on the constructor. And change the add to pre set
numbers which can contain turn instance variable
play_roll:
changed the name of the method, previusly roll, to avoid confusion with
instance variable with the same name